Create CaseMailer #3200
