+++ app/controllers/cases_controller.rb
@@ -64,7 +64,7 @@ class CasesController < ApplicationController
     if @this_case.update_attributes(case_params)
       flash[:success] = 'Case was updated!'
       flash[:undo] = @this_case.versions
-      UserNotifier.send_followers_email(@this_case.followers, @this_case).deliver_now
+      CaseMailer.send_followers_email(@this_case.followers, @this_case).deliver_now
       redirect_to @this_case
     else
       set_instance_vars
@@ -78,7 +78,7 @@ class CasesController < ApplicationController
       @this_case.destroy
       flash[:success] = 'Case was removed!'
       flash[:undo] = @this_case.versions
-      UserNotifier.send_deletion_email(@this_case.followers, @this_case).deliver_now
+      CaseMailer.send_deletion_email(@this_case.followers, @this_case).deliver_now
     rescue ActiveRecord::RecordNotFound
       flash[:notice] = I18n.t('cases_controller.case_not_found_message')
     end
