+++ app/mailers/case_mailer.rb
@@ -1,0 +1,19 @@
+class CaseMailer < ApplicationMailer
+  default from: 'EndBiasWiki@gmail.com'
+ 
+ def send_followers_email(users, this_case)
+    @this_case = this_case
+    users.each do |user|
+      Rails.logger.info("UserNotifier#send_followers_email: Sending notification to #{user.email} about case #{this_case.title}")
+      mail(to: user.email, subject: "The #{@this_case.title} case has been updated on EBWiki.")
+    end
+  end
+
+  def send_deletion_email(users, this_case)
+    @this_case = this_case
+    users.each do |user|
+      Rails.logger.info("UserNotifier#send_deletion_email: Sending notification to #{user.email} about case #{this_case.title}")
+      mail(to: user.email, subject: "The #{@this_case.title} case has been removed from EBWiki")
+    end
+  end
+end
