+++ app/views/case_mailer/send_followers_email.html.erb
